// backend/src/lib/api-zod/src/generated/api.ts
export const CreateTaskBody = zod.object({
  title: zod.string(),
  description: zod.string(),
  taskType: zod.enum([
    "physical",
    "mental",
    "creative",
    "romantic",
    "elimination",
  ]),
  scheduledAt: zod.date().nullish(),
});

/**
 * @summary Update task result
 */
export const UpdateTaskParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateTaskBody = zod.object({
  status: zod.enum(["upcoming", "ongoing", "completed"]),
  winnerIds: zod.string().nullish(),
});
// ...
export const DeleteTaskParams = zod.object({
  id: zod.coerce.number(),
});